ai-engine/requirements.txt
python-docx>=1.1.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
qdrant-client>=1.7.0
requests>=2.31.0
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
langchain-core>=0.1.0
typing-extensions>=4.8.0
tavily-python>=0.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
